refactor: put all functions for filtering in one file
