+++ src/utils/general.ts
@@ -40,27 +40,9 @@ const extractTweetAuthor = (tweet: HTMLDivElement): string => {
     return tweetAuthor;
 };
 
-const isAccountPrivate = (tweet: HTMLDivElement): boolean => {
-    return (
-        tweet.querySelectorAll('[aria-label="Protected account"]').length > 0
-    );
-};
-
-const isPostedByCurrentUser = (
-    tweet: HTMLDivElement,
-    currentUser: string
-): boolean => {
-    const tweetContent = tweet.innerText.split("\n");
-    const tweetAuthor =
-        tweetContent[tweetContent[0].includes("reposted") ? 2 : 1];
-    return currentUser === tweetAuthor;
-};
-
 export {
     getCurrentTab,
     getCurrentUsername,
-    isAccountPrivate,
-    isPostedByCurrentUser,
     sendExtensionStateToContentScript,
     extractTweetAuthor,
 };
