+++ src/utils/index.ts
@@ -3,8 +3,6 @@ import {
     getCurrentTab,
     getCurrentUsername,
     sendExtensionStateToContentScript,
-    isAccountPrivate,
-    isPostedByCurrentUser,
     extractTweetAuthor,
 } from "./general";
 import {
@@ -13,7 +11,11 @@ import {
     setStoredAllTimeTweetCount,
     getStoredAllTimeTweetCount,
 } from "./storage";
-import { isFromNewsOutlet } from "./filters";
+import {
+    isFromNewsOutlet,
+    isAccountPrivate,
+    isPostedByCurrentUser,
+} from "./filters";
 import { mediaDirectory } from "./media-directory";
 
 export {
